Feature/speedometer colors (#6)
* Add colors to speedometer based on acceleration

Added colors to the speedometer based on acceleration. The coloring
algorithm works by taking the difference in speeds from the previous
frame and current frame to calculate acceleration. This value is then
smoothed with a simple lowpass function. Then it is "compressed" into a
value from 0.0 to 1.0. (This value can "saturate" and is currently
implemented with an atan function.) This new value is used to
interpolate between the neutral color and either the positive
acceleration color or the negative acceleration color.

The code has been split out into its own class file since it had its own
separate state.

The gameMainVehicleSpeed global value was also changed from an int to a
float to enable better acceleration calculations.

* Simplify accel compression using clamp

* Use scaleCopy in GUINumber